+++ MeldingerReceiver/AT.Common.MeldingerReceiver.Publish/Abstract/MeldingerReceiverBase.cs
@@ -1,6 +1,0 @@
-namespace Arbeidstilsynet.Common.MeldingerReceiver.Abstract;
-
-public abstract class MeldingerReceiverBase
-{
-    public abstract string Foo { get; }
-}