+++ MeldingerReceiver/AT.Common.MeldingerReceiver.Publish/DependencyInjection/DependencyInjectionExtensions.cs
@@ -1,7 +1,13 @@
 using Arbeidstilsynet.Common.MeldingerReceiver.Implementation;
 using Microsoft.Extensions.DependencyInjection;
+using StackExchange.Redis;
 
 namespace Arbeidstilsynet.Common.MeldingerReceiver.DependencyInjection;
+
+public record ValkeyConfiguration
+{
+    public required string ConnectionString { get; init; } = "localhost";
+}
 
 /// <summary>
 /// Extensions for Dependency Injection.
@@ -12,10 +18,12 @@ public static class DependencyInjectionExtensions
     /// Registrerer en implementasjon av IMeldingerReceiver i den spesifiserte <see cref="IServiceCollection"/>.
     /// </summary>
     /// <param name="services"><see cref="IServiceCollection"/> som tjenesten skal legges til i.</param>
+    /// <param name="configuration"></param>
     /// <returns><see cref="IServiceCollection"/> for chaining.</returns>
-    public static IServiceCollection AddMeldingerReceiver(this IServiceCollection services)
+    public static IServiceCollection AddMeldingerReceiver(this IServiceCollection services, ValkeyConfiguration configuration)
     {
-        services.AddSingleton<IMeldingerReceiver, MeldingerReceiverImplementation>();
+        services.AddSingleton<IConnectionMultiplexer>(ConnectionMultiplexer.Connect(configuration.ConnectionString));
+        services.AddSingleton<IMeldingerReceiver, Implementation.MeldingerReceiver>();
 
         return services;
     }