+++ MeldingerReceiver/AT.Common.MeldingerReceiver.Publish/Ports/IConstants.cs
@@ -1,0 +1,9 @@
+namespace Arbeidstilsynet.Common.MeldingerReceiver;
+
+public interface IConstants
+{
+
+    public const string MessageKey = "melding-received";
+    
+    public const string StreamName = "meldinger-receiver";
+}